+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.3] - Unreleased
 
+### Fixed
+
+- Fix duplicate event creation in `scan` method when status and verification state are unchanged
+
 ## [0.2.2] - 2025-09-04
 
 ### Changed